+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 0.0.3 (August 28th, 2010)
+* Fixed misspelling of "repository_path" in Mysql2Model::Config
+* Added specs for all classes
+* Added value method to Mysql2Model::Container
+* Added interpolating values into the queries called "composing"
 ## 0.0.2 (August 26th, 2010)
 * First Working Version
 ## 0.0.1 (August 25th, 2010)